backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
websockets==12.0
python-multipart==0.0.6